Blog ownership - view replies for users' blog
- Github authentication
- Create social_login and blog_ownership tables
- Draft page for analyzing blog stats
- Removal of unsafe endpoints
- Code cleanup

+++ src/main/java/me/mrabar/sq/model/Blog.java
@@ -1,4 +1,4 @@
 package me.mrabar.sq.model;
 
-public record Blog(long id, String name, String key, BlogMisc misc) {
+public record Blog(int id, String name, String key, BlogMisc misc) {
 }
